Extract loop data manager dosing tests into their own file
